+++ package.json
@@ -15,6 +15,7 @@
     "typescript": "^5.4.5"
   },
   "dependencies": {
+    "dotenv": "^16.4.5",
     "envalid": "^8.0.0",
     "playwright": "^1.48.0",
     "swag-labs": "file:"
