+++ playwright.config.ts
@@ -1,4 +1,6 @@
 import { defineConfig, devices } from '@playwright/test'
+import * as dotenv from 'dotenv';
+dotenv.config();
 // import path from 'path'
 /**
  * Read environment variables from file.
@@ -16,7 +18,7 @@ export default defineConfig({
   /* Run tests in files in parallel */
   fullyParallel: false,
   /* Fail the build on CI if you accidentally left test.only in the source code. */
-  //forbidOnly: !!process.env.CI,
+  // forbidOnly: !!process.env.CI,
   /* Retry on CI only */
   //retries: process.env.CI ? 2 : 0,
   /* Opt out of parallel tests on CI. */
@@ -37,7 +39,7 @@ export default defineConfig({
     //baseURL: 'https://example.com',
 
     /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
-    headless: true
+    headless: false
   },
   // General timeout per test
   timeout: 20000,
@@ -49,6 +51,17 @@ export default defineConfig({
       viewport: null
       },
     }
+
+    // Uncomment browser you want to run
+    // {
+    //   name: 'firefox',
+    //   use: { ...devices['Desktop Firefox'] },
+    // },
+
+    // {
+    //   name: 'webkit',
+    //   use: { ...devices['Desktop Safari'] },
+    // },
 
   ],
